requirements.txt
amqp
appdirs
astroid
billiard==3.6.3.0
cached-property==1.5.2
celery
certifi
chardet
click
click-didyoumean
click-plugins
click-repl
cycler==0.10.0
defusedxml==0.5.0
future==0.18.2
idna==2.8
importlib-metadata
isodate
isort==5.7.0
kiwisolver
kombu
kombu-sqlalchemy
lazy-object-proxy
lxml
matplotlib
mccabe==0.6.1
numpy
pandas
prompt-toolkit==3.0.10
pylint==2.15.3
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.13.0
python-telegram-bot==9.0.0
pytz
requests==2.22.0
requests-toolbelt==0.9.1
six
SQLAlchemy
toml
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
xlrd
zeep==2.5.0
zipp==3.8.1
